# Optimize CI times