--- a/tests/out/hlsl/access.hlsl.config
+++ b/tests/out/hlsl/access.hlsl.config
@@ -1,3 +1,0 @@
-vertex=(foo_vert:vs_5_1 )
-fragment=(foo_frag:ps_5_1 )
-compute=(assign_through_ptr:cs_5_1 )