--- a/tests/out/hlsl/array-in-ctor.hlsl.config
+++ b/tests/out/hlsl/array-in-ctor.hlsl.config
@@ -1,3 +1,0 @@
-vertex=()
-fragment=()
-compute=(cs_main:cs_5_1 )