--- a/tests/out/hlsl/array-in-ctor.ron
+++ b/tests/out/hlsl/array-in-ctor.ron
@@ -1,0 +1,12 @@
+(
+    vertex:[
+    ],
+    fragment:[
+    ],
+    compute:[
+        (
+            entry_point:"cs_main",
+            target_profile:"cs_5_1",
+        ),
+    ],
+)