--- a/tests/out/hlsl/atomicOps.hlsl.config
+++ b/tests/out/hlsl/atomicOps.hlsl.config
@@ -1,3 +1,0 @@
-vertex=()
-fragment=()
-compute=(cs_main:cs_5_1 )