--- a/tests/out/hlsl/atomicOps.ron
+++ b/tests/out/hlsl/atomicOps.ron
@@ -1,0 +1,12 @@
+(
+    vertex:[
+    ],
+    fragment:[
+    ],
+    compute:[
+        (
+            entry_point:"cs_main",
+            target_profile:"cs_5_1",
+        ),
+    ],
+)