--- a/tests/out/hlsl/binding-arrays.hlsl.config
+++ b/tests/out/hlsl/binding-arrays.hlsl.config
@@ -1,3 +1,0 @@
-vertex=()
-fragment=(main:ps_5_1 )
-compute=()