--- a/tests/out/hlsl/binding-arrays.ron
+++ b/tests/out/hlsl/binding-arrays.ron
@@ -1,0 +1,12 @@
+(
+    vertex:[
+    ],
+    fragment:[
+        (
+            entry_point:"main",
+            target_profile:"ps_5_1",
+        ),
+    ],
+    compute:[
+    ],
+)