--- a/tests/out/hlsl/bitcast.hlsl.config
+++ b/tests/out/hlsl/bitcast.hlsl.config
@@ -1,3 +1,0 @@
-vertex=()
-fragment=()
-compute=(main:cs_5_1 )