--- a/xtask/.gitignore
+++ b/xtask/.gitignore
@@ -1,0 +1,2 @@
+!Cargo.lock
+target/