| File, job name, optionally step     | Total runtime of job (pre-caching) | Total runtime of step (pre-caching) | Time spent compiling stuff with Cargo (pre-caching)                                                       | Time spent compiling stuff with Cargo (post-caching)                                                | Cache size                                                                                               |
| ----------------------------------- | ---------------------------------- | ----------------------------------- | ------------------------------------------------------------------------------------------- | ----------------------------------------------------------------------------------------------------- | -------------------------------------------------------------------------------------------------------- |
| `ci.yml:check-msrv`                 | 2m 16s                             | -                                   | [1m 43s](https://github.com/gfx-rs/naga/actions/runs/4606099344/jobs/8138976872#step:7:98)  | [1m 0s](https://github.com/ErichDonGubler/naga/actions/runs/4612139310/jobs/8152686869?pr=53)         | [139 MB](https://github.com/ErichDonGubler/naga/actions/runs/4612577974/jobs/8153632565#step:7:42)       |
| `ci.yml:test:Default test`          | 4m 12s                             | 2m 07s                              | [2m 07s](https://github.com/gfx-rs/naga/actions/runs/4606099344/jobs/8138977360#step:4:145) | [17s](https://github.com/ErichDonGubler/naga/actions/runs/4612139310/jobs/8152687504?pr=53#step:5:13) | [303 MB](https://github.com/ErichDonGubler/naga/actions/runs/4612577974/jobs/8153633048#step:4:42)       |
| `ci.yml:test:Test all features`     | 4m 12s                             | 1m 53s                              | [1m 39s](https://github.com/gfx-rs/naga/actions/runs/4606099344/jobs/8138977360#step:5:96)  | [46s](https://github.com/ErichDonGubler/naga/actions/runs/4612139310/jobs/8152687504?pr=53#step:5:13) | [303 MB](https://github.com/ErichDonGubler/naga/actions/runs/4612577974/jobs/8153633048#step:4:42)       |
| `ci.yml:check:Check benchmarks`     | 2m 49s                             | 2m 09s                              | [2m 08s](https://github.com/gfx-rs/naga/actions/runs/4606099344/jobs/8138976718#step:3:166) | [19s](https://github.com/ErichDonGubler/naga/actions/runs/4612577974/jobs/8153632686?pr=53#step:4:14) | [146 MB](https://github.com/ErichDonGubler/naga/actions/runs/4612577974/jobs/8153632686)                 |
| `ci.yml:check:Check naga-fuzz`      | 2m 49s                             | 36s                                 | [35s](https://github.com/gfx-rs/naga/actions/runs/4606099344/jobs/8138976718#step:4:54)     | [30s](https://github.com/ErichDonGubler/naga/actions/runs/4612577974/jobs/8153632686?pr=53#step:5:61) | [146 MB](https://github.com/ErichDonGubler/naga/actions/runs/4612577974/jobs/8153632686?pr=53#step:3:42) |
| `ci.yml:documentation`              | 1m 59s                             | -                                   | [1m 55s](https://github.com/gfx-rs/naga/actions/runs/4606099344/jobs/8138977048#step:3:103) | [25s](https://github.com/ErichDonGubler/naga/actions/runs/4612139310/jobs/8152687375?pr=53#step:4:40) | [127 MB](https://github.com/ErichDonGubler/naga/actions/runs/4612139310/jobs/8152687375?pr=53#step:3:43) |
| `validate-linux.yml:validate-linux` | 2m 29s                             | -                                   | [2m 06s](https://github.com/gfx-rs/naga/actions/runs/4606099371/jobs/8138976708#step:7:92)  |                                                                                                       |                                                                                                          |